+++ dflib-echarts/src/main/java/org/dflib/echarts/CartesianSeriesOpts.java
@@ -6,6 +6,7 @@ package org.dflib.echarts;
 public abstract class CartesianSeriesOpts<T extends CartesianSeriesOpts<T>> extends SeriesOpts {
 
     protected Label label;
+    protected Integer xAxisIndex;
     protected Integer yAxisIndex;
 
     /**
@@ -21,6 +22,17 @@ public abstract class CartesianSeriesOpts<T extends CartesianSeriesOpts<T>> exte
      */
     public T label(Label label) {
         this.label = label;
+        return (T) this;
+    }
+
+    /**
+     * Sets an index of X axis to use for this Series. There can be one or more X axes, so this method allows to
+     * pick one. If not set, 0 is assumed.
+     *
+     * @since 1.0.0-M22
+     */
+    public T xAxisIndex(int index) {
+        this.xAxisIndex = index;
         return (T) this;
     }