+++ dflib-echarts/src/main/java/org/dflib/echarts/LineSeriesOpts.java
@@ -48,6 +48,7 @@ public class LineSeriesOpts extends CartesianSeriesOpts<LineSeriesOpts> {
                 showSymbol,
                 stack,
                 smooth,
+                xAxisIndex,
                 yAxisIndex,
                 null,
                 null,