+++ dflib-echarts/src/main/java/org/dflib/echarts/ScatterSeriesOpts.java
@@ -20,6 +20,7 @@ public class ScatterSeriesOpts extends CartesianSeriesOpts<ScatterSeriesOpts> {
                 null,
                 null,
                 null,
+                xAxisIndex,
                 yAxisIndex,
                 null,
                 null,