+++ dflib-echarts/src/main/java/org/dflib/echarts/render/option/axis/AxisLineModel.java
@@ -4,13 +4,24 @@ package org.dflib.echarts.render.option.axis;
  * @since 1.0.0-M22
  */
 public class AxisLineModel {
-    private final boolean show;
 
-    public AxisLineModel(boolean show) {
+    private final boolean show;
+    private final Boolean onZero;
+
+    public AxisLineModel(boolean show, Boolean onZero) {
         this.show = show;
+        this.onZero = onZero;
     }
 
     public boolean isShow() {
         return show;
+    }
+
+    public boolean isOnZeroPresent() {
+        return onZero != null;
+    }
+
+    public Boolean getOnZero() {
+        return onZero;
     }
 }